[Breaking] Add `config` feature for `sailfish` crate
This is for enabling the corresponding `config` feature in
`sailfish-macros` and `sailfish-compiler`. If it is disabled,
configuration files (sailfish.toml) will be ignored. More than 10 cargo
build steps can be avoided this way, since the `toml` and `serde`
dependencies in `sailfish-compiler` are no longer needed.

This change will break applications that use sailfish.toml and
`default-features = false` for the `sailfish` dependency. They need to
add `config` to the list of enabled `sailfish` features in Cargo.toml.

Closes #123