Cache templates compiled by proc-macro
Instead of recompiling sailfish templates on every proc-macro
invocation, see if an output file for the same input file content +
compiler config combination already exists.

This also fixes #58 because it avoids multiple proc-macro invocations
writing to the same output file at roughly the same time from different
processes, for example in clippy checks that execute in parallel.

In addition to the compiled output, now the list of dependencies (file
names), which was previously generated by the compiler for each template
on every proc-macro invocation, needs to be stored to disk for re-use.